Boldi's first fommit
The second half of `2_methods.tex` has been formatted, and the required elements have been created.

Also, `\begin{document}...\end{document}` has been added to each file to include the content of them.
Based on: https://www.overleaf.com/learn/latex/Multi-file_LaTeX_projects

Lastly, `\section{Összefoglalás}` has become a `\section*` since I believe that it does not have to be a real section.
If it is not thought to be a good idea, it can be undone.